auth_service/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
# PostgreSQL
psycopg2-binary==2.9.6

python-dotenv==1.0.0
django-environ==0.11.2

# Redis + Celery
redis==4.5.4
celery[redis]==5.3.4

# RabbitMQ
pika==1.3.2

# Google Login
social-auth-app-django==5.4.0
social-auth-core==4.5.0
requests==2.31.0

# S3
boto3==1.28.0
django-storages==1.14.2

# Images
Pillow==10.2.0
google-auth
google-auth-oauthlib